# src/config/swagger2.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bookstore API",
    "version": "1.0.0",
    "description": "API documentation for book-related operations"
  },
...
  "paths": {
...
    "/api/bookstore/books/": {
      "get": {
        "tags": ["Book"],
        "summary": "Retrieve all books",
        "description": "Get all books from the bookstore",
        "responses": {
          "200": {
            "description": "A list of books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Book"],
        "summary": "Create a new book",
        "description": "Add a new book to the bookstore",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookstore/books/{id}": {
      "get": {
        "tags": ["Book"],
        "summary": "Get a book by ID",
        "description": "Get a book from the bookstore by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID of the book to retrieve",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single book",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Book"],
        "summary": "Update a book by ID",
        "description": "Update a book in the bookstore by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID of the book to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Book"],
        "summary": "Delete a book by ID",
        "description": "Delete a book from the bookstore by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID of the book to delete",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the book.",
            "example": "Electronics Tutorial"
          },
          "writer": {
            "type": "string",
            "description": "The name of the writer/author of the book.",
            "example": "Carlos Harber"
          },
          "cover": {
            "type": "string",
            "format": "url",
            "description": "URL to the book's cover image.",
            "example": "https://images-na.ssl-images-amazon.com/images/I/51Ga5GuElyL._AC_SX184_.jpg"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "The price of the book.",
            "example": 25
          },
          "tag": {
            "type": "array",
            "items": {
              "type": "string",
              "example": ["fiction"]
            },
            "nullable": true,
            "description": "Tags associated with the book."
          },
          "description": {
            "type": "string",
            "description": "A brief description of the book.",
            "example": "Corrumpo curto bos paulatim. Quos arcesso centum vulgivagus uter alter. Verto admoneo repellat thymbra vigilo nulla."
          }
        }
      },
...
    }
  }
}